Add new sign in screen and folder structure
