+++ app/src/main/java/utn/frba/mobile/moodmatch/ui/theme/Theme.kt
@@ -1,6 +1,5 @@
 package utn.frba.mobile.moodmatch.ui.theme
 
-import android.app.Activity
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
